Added Controllers: save, getFileLocation, Serialize
There is now a controller that handles pushing the save button (in the
menu bar only for now), a separate controller for getting a file
location, and the last controller for serializing the LevelTemplate and
writing to file. I broke them down this way so they can be easily
reused. assumes there is a model class that lets them get a reference to
the LevelTemplate object.